requirements.txt
discord.py>=2.3.0
flask>=2.0.0
flask-sqlalchemy>=3.0.0
aiohttp>=3.8.0
python-dotenv==0.19.0
psycopg2-binary
authlib
supabase